services/etl/requirements.txt
pandas
numpy
psycopg2-binary
python-dotenv
requests
yfinance
sqlalchemy
schedule
